pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-cli-agent"
version = "0.1.0"
description = "A natural language powered CLI agent for local tasks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.60",
    "anthropic>=0.42",
    "click>=8.1",
    "rich>=13.7",
    "python-dotenv>=1.0",
]

[project.scripts]
local-cli-agent = "local_cli_agent.main:main"

[tool.setuptools.packages.find]
include = ["local_cli_agent*"]
